move common functioality in the create-account and add-workplace type of employer component into a directive
